added better closing func
in which key - not a fan of it's palcement. need to have a utils or custom functions file. idk maybe having the samll func where it does it's job isn't a bad idea.
now on <leader>q i bdelete if the file is saved and named and force q! if new. helps while
I am iterating on files.